Moved include/IWallet.h to lib/Wallet/IWallet.h
