+++ lib/Wallet/IWallet.h
